requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
google-api-core==2.25.1
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-storage==3.3.0
    # via my-video-app (pyproject.toml)
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via google-api-core
idna==3.10
    # via requests
proto-plus==1.26.1
    # via google-api-core
protobuf==6.32.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
requests==2.32.5
    # via
    #   google-api-core
    #   google-cloud-storage
rsa==4.9.1
    # via google-auth
urllib3==2.5.0
    # via requests
